Move polycube code into polycube.rs
